+++ data/translations/optimus-manager_de.ts
@@ -363,8 +363,8 @@ Warnung ignorieren und mit GPU-Wechsel fortfahren?</translation>
     </message>
     <message>
         <location filename="../../src/settingsdialog.ui" line="398"/>
-        <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Automatically log out the current desktop session when switching GPUs&lt;/p&gt;&lt;p&gt;At the moment this is only supported for Gnome, KDE Plasma, Deepin, Xfce, i3 and Sway.&lt;/p&gt;&lt;p&gt;If this options is disabled or you use a different desktop environment, GPU switching only becomes effective at the next graphical session login&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
-        <translation>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Aktuelle Desktop-Sitzung bei GPU-Wechsel automatisch ausloggen&lt;/p&gt;&lt;p&gt;Zurzeit nur unter Gnome, KDE Plasma, Deepin, Xfce, i3 und Sway ünterstützt.&lt;/p&gt;&lt;p&gt;Ist diese Option deaktiviert oder eine andere Desktopumgebung wird verwendet, wird der GPU-Wechsel erst beim nächsten manuellen Logout der Sitzung aktiv.&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</translation>
+        <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Automatically log out the current desktop session when switching GPUs&lt;/p&gt;&lt;p&gt;At the moment this is only supported for Gnome, KDE Plasma, Deepin, Xfce, i3, Sway and Openbox&lt;/p&gt;&lt;p&gt;If this options is disabled or you use a different desktop environment, GPU switching only becomes effective at the next graphical session login&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
+        <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../../src/settingsdialog.ui" line="444"/>